refactor: Rename `db_adapters` -> `adapters`
