+++ database/src/adapters/assets/event_types.rs
