+++ database/src/adapters/assets/mod.rs
