+++ database/src/adapters/execution_outcomes.rs
