+++ database/src/adapters/mod.rs
