+++ database/src/adapters/receipts.rs
